+++ packages/client/src/App.tsx
@@ -22,6 +22,7 @@ import StockArticlesPage from "./pages/StockArticlesPage";
 import CommissionsPage from "./pages/CommissionsPage";
 import CommissionEditQueuePage from "./pages/CommissionEditQueuePage";
 import DashboardPage from "./pages/DashboardPage";
+import EmployeeActivityDetailPage from "./pages/EmployeeActivityDetailPage";
 
 interface AuthUser {
   id: number;
@@ -136,6 +137,7 @@ export default function App() {
             }
           >
             <Route path="/dashboard" element={<DashboardPage />} />
+            <Route path="/reporting/employees/:agentId" element={<EmployeeActivityDetailPage />} />
             <Route path="/parties" element={<PartiesPage />} />
             <Route path="/parties/:id" element={<PartyDetailPage />} />
             <Route path="/proformas" element={<ProformasPage />} />